broken down ImmutableHelpers.test into it's sub components so it matchs with the way we handle src folder
